Conflicts:
	Applications/MIDIMonitor/English.lproj/MainMenu.nib/designable.nib
	Applications/MIDIMonitor/English.lproj/MainMenu.nib/keyedobjects.nib
	Frameworks/SnoizeMIDI/SMInputStream.m
	Frameworks/SnoizeMIDI/SMMIDIObject.m
	Frameworks/SnoizeMIDI/SnoizeMIDI.xcodeproj/project.pbxproj
	Frameworks/SnoizeMIDI/SnoizeMIDI.xcodeproj/xcshareddata/xcschemes/SnoizeMIDI.xcscheme
	Frameworks/SnoizeMIDI/SnoizeMIDI.xcodeproj/xcshareddata/xcschemes/SnoizeMIDIiOS.xcscheme

+++ Frameworks/SnoizeMIDISpy/Framework/MIDISpyClient.c
@@ -154,7 +154,7 @@ OSStatus MIDISpyClientCreate(MIDISpyClientRef *outClientRefPtr)
 
         // Now get the identifier and use it to name a newly created local port
         clientRef->clientIdentifier = *(UInt32 *)CFDataGetBytePtr(identifierData);
-        localPortName = CFStringCreateWithFormat(kCFAllocatorDefault, NULL, CFSTR("%@-%lu"), kSpyingMIDIDriverPortName, clientRef->clientIdentifier);
+        localPortName = CFStringCreateWithFormat(kCFAllocatorDefault, NULL, CFSTR("%@-%u"), kSpyingMIDIDriverPortName, (unsigned int)clientRef->clientIdentifier);
 
         context.info = clientRef;
         clientRef->localPort = CFMessagePortCreateLocal(kCFAllocatorDefault, localPortName, LocalMessagePortCallback, &context, FALSE);
